\name{ABC4ESD}
\alias{ABC4ESD}
\title{ABC for Empirical-Statistical Downscaling.}
\description{Downloads the compendium 'Empirical-Statistical
  Downscaling' from \url{http://home.broadpark.no/~rbene}. Other literature on matters relating to clim.pact can be found on \url{http://home.broadpark.no/~rbene/ras-publ.html}
}
\usage{
ABC4ESD(browser="firefox")
}
\arguments{
  \item{browser}{Browser to use}
 }
\value{a PDF-document.}
\author{R.E. Benestad}
\examples{
}
\keyword{}
