\name{lonlat2utm}
\alias{lonlat2utm}
\title{Convert from Longitude-Latitude to UTM}
\description{NOT FINISHED YET: Temporary version only
  approximate. Assumes a perfect sphere.}
\usage{lonlat2utm(lon,lat,a=6.378e06)}
\arguments{
    \item{lon}{longitude in degrees east.}
    \item{lat}{latitude in degrees north.}
    \item{a}{radius of Earth.}
 }
\value{list}
\author{R.E. Benestad}}
\examples{
}
\keyword{manip}