\name{newFig}
\alias{newFig}
\title{Create a new Window}
\description{Creates a new window by calling the appropriate device, eg
  \code{\link{x11}}(), \code{windows()}, \code{\link{pdf}}(),
  \code{\link{postscript}}(),etc.}
\usage{newFig()}
\author{R.E. Benestad}
\examples{
newFig()
}
\keyword{device}
