\name{utm2lonlat}
\alias{utm2lonlat}
\title{Convert from UTM to Longitude-Latitude}
\description{NOT FINISHED YET: Temporary version only
  approximate. Assumes a perfect sphere.}
\usage{utm2lonlat(easting,northing,zone,band=NULL,a=6.378e06)}
\arguments{
    \item{easting}{.}
    \item{northing}{.}
    \item{zone}{.}
    \item{band}{.}
    \item{a}{radius of Earth.}
 }
\value{list}
\author{R.E. Benestad}}
\examples{
}
\keyword{manip}