\name{ABC4ESD}
\alias{ABC4ESD}
\title{ABC for Empirical-Statistical Downscaling.}
\description{Downloads the compendium 'Empirical-Statistical
  Downscaling' from \url{http://home.broadpark.no/~rbene}. Other literature on matters relating to clim.pact can be found on \url{http://home.broadpark.no/~rbene/ras-publ.html}
}
\usage{
ABC4ESD(browser="firefox",url="http://rcg.gvc.gu.se/edu/esd.pdf")
}
\arguments{
  \item{browser}{Browser to use}
  \item{url}{URL for the ESD-compendium.}
 }
 \value{a PDF-document.}
 \references{Benestad, R.E., Hanssen-Bauer, I. And Chen, D.(2008)
   Empirical-Statistical Downscaling, World Scientific Publishers, ISBN
   978-981-281-912-3 \url{http://www.worldscibooks.com/environsci/6908.html}}
\author{R.E. Benestad}
\examples{
\dontrun{ABC4ESD()}
}
\keyword{methods}
